Don't skip the full test when we're filtering out a subtest
This showed up with anodyne because they were failing the pickle dumps
but it wasn't showing with this script. Turns out we cannot prevent
subtests from executing and calling skipTest was exiting the parent test
completely. So we now suppress all exceptions for subtests we ignore
(which is fine, it should just be ALTTP since all other worlds get
unloaded)

Fixes #4